// libs/enums/notification.enum.ts
export enum NotificationType {
	LIKE = 'LIKE',
	COMMENT = 'COMMENT',
	MEMBER_FOLLOW = 'MEMBER_FOLLOW',
	PRODUCT_INQUIRY = 'PRODUCT_INQUIRY',
	PRODUCT_INQUIRY_RECEIVED = 'PRODUCT_INQUIRY_RECEIVED',
	PRODUCT_INQUIRY_RESPONDED = 'PRODUCT_INQUIRY_RESPONDED',
	EVENT_BOOKING_RECEIVED = 'EVENT_BOOKING_RECEIVED',
	EVENT_BOOKING_CONFIRMED = 'EVENT_BOOKING_CONFIRMED',
	EVENT_BOOKING_CANCELLED = 'EVENT_BOOKING_CANCELLED',
}

export enum NotificationStatus {
	WAIT = 'WAIT',
	READ = 'READ',
}

export enum NotificationGroup {
	MEMBER = 'MEMBER',
	EVENT = 'EVENT',
	PRODUCT = 'PRODUCT',
	INQUIRY = 'INQUIRY',
	BOOKING = 'BOOKING',
}